+++ src/Splat.Autofac.Tests/Splat.Autofac.Tests.csproj
@@ -10,11 +10,11 @@
 
   <ItemGroup>
     <PackageReference Include="Autofac" Version="4.8.1" />
-    <PackageReference Include="ReactiveUI" Version="*" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Splat.Autofac\Splat.Autofac.csproj" />
+    <ProjectReference Include="..\Splat.Common.Testing\Splat.Common.Testing.csproj" />
   </ItemGroup>
   
   <ItemGroup>
