+++ src/Splat.Common.Testing/ViewModelOne.cs
@@ -3,21 +3,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
-using ReactiveUI;
-
-namespace Splat.Autofac.Tests
+namespace Splat.Common.Testing
 {
     /// <summary>
     /// View Model One.
     /// </summary>
-    /// <seealso cref="ReactiveObject" />
-    /// <seealso cref="IRoutableViewModel" />
-    public class ViewModelOne : ReactiveObject, IRoutableViewModel
+    public class ViewModelOne
     {
-        /// <inheritdoc />
-        public string UrlPathSegment { get; }
-
-        /// <inheritdoc />
-        public IScreen HostScreen { get; }
     }
 }
