+++ src/Splat.Common.Testing/ViewOne.cs
@@ -3,9 +3,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
-using ReactiveUI;
-
-namespace Splat.DryIoc.Tests
+namespace Splat.Common.Testing
 {
     /// <summary>
     /// View One.
