+++ src/Splat.DryIoc.Tests/DependencyResolverTests.cs
@@ -7,8 +7,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using DryIoc;
-using ReactiveUI;
 using Shouldly;
+using Splat.Common.Testing;
 using Xunit;
 
 namespace Splat.DryIoc.Tests
@@ -86,42 +86,6 @@ namespace Splat.DryIoc.Tests
 
             screen.ShouldNotBeNull();
             screen.ShouldBeOfType<MockScreen>();
-        }
-
-        /// <summary>
-        /// Shoulds register ReactiveUI binding type converters.
-        /// </summary>
-        [Fact]
-        public void DryIocDependencyResolver_Should_Register_ReactiveUI_BindingTypeConverters()
-        {
-            // Invoke RxApp which initializes the ReactiveUI platform.
-            var scheduler = RxApp.MainThreadScheduler;
-            var container = new Container();
-            container.UseDryIocDependencyResolver();
-
-            var converters = container.Resolve<IEnumerable<IBindingTypeConverter>>().ToList();
-
-            converters.ShouldNotBeNull();
-            converters.ShouldContain(x => x.GetType() == typeof(StringConverter));
-            converters.ShouldContain(x => x.GetType() == typeof(EqualityTypeConverter));
-        }
-
-        /// <summary>
-        /// Shoulds register ReactiveUI creates command bindings.
-        /// </summary>
-        [Fact]
-        public void DryIocDependencyResolver_Should_Register_ReactiveUI_CreatesCommandBinding()
-        {
-            // Invoke RxApp which initializes the ReactiveUI platform.
-            var scheduler = RxApp.MainThreadScheduler;
-            var container = new Container();
-            container.UseDryIocDependencyResolver();
-
-            var converters = container.Resolve<IEnumerable<ICreatesCommandBinding>>().ToList();
-
-            converters.ShouldNotBeNull();
-            converters.ShouldContain(x => x.GetType() == typeof(CreatesCommandBindingViaEvent));
-            converters.ShouldContain(x => x.GetType() == typeof(CreatesCommandBindingViaCommandParameter));
         }
     }
 }
