+++ src/Splat.SimpleInjector.Tests/Splat.SimpleInjector.Tests.csproj
@@ -7,10 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="ReactiveUI" Version="*" />
-  </ItemGroup>
-
-  <ItemGroup>
+    <ProjectReference Include="..\Splat.Common.Testing\Splat.Common.Testing.csproj" />
     <ProjectReference Include="..\Splat.SimpleInjector\Splat.SimpleInjector.csproj" />
   </ItemGroup>
