--- src/Splat.SimpleInjector.Tests/ViewModel.cs
@@ -1,14 +1,0 @@
-// Copyright (c) 2019 .NET Foundation and Contributors. All rights reserved.
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for full license information.
-
-namespace Splat.Simplnjector
-{
-    /// <summary>
-    /// View Model.
-    /// </summary>
-    public class ViewModel
-    {
-    }
-}
